update readme file, various small fixs

+++ backend/src/products/products.service.ts
@@ -1,31 +1,48 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { PrismaService } from '../prisma/prisma.service';
+import { ProductEntity } from './entities/product.entity';
 
 @Injectable()
 export class ProductsService {
   constructor(private prisma: PrismaService) {}
-  create(createProductDto: CreateProductDto) {
+
+  async create(createProductDto: CreateProductDto): Promise<ProductEntity> {
     return this.prisma.product.create({ data: createProductDto });
   }
 
-  findAll() {
+  async findAll(): Promise<ProductEntity[]> {
     return this.prisma.product.findMany();
   }
 
-  findOne(id: number) {
-    return this.prisma.product.findUnique({ where: { id } });
+  async findOne(id: number): Promise<ProductEntity> {
+    const product = await this.prisma.product.findUnique({ where: { id } });
+    if (!product) {
+      throw new NotFoundException(`Product with id ${id} does not exist`);
+    }
+    return product;
   }
 
-  update(id: number, updateProductDto: UpdateProductDto) {
+  async update(
+    id: number,
+    updateProductDto: UpdateProductDto,
+  ): Promise<ProductEntity> {
+    const product = await this.prisma.product.findUnique({ where: { id } });
+    if (!product) {
+      throw new NotFoundException(`Product with id ${id} does not exist`);
+    }
     return this.prisma.product.update({
       where: { id },
       data: updateProductDto,
     });
   }
 
-  remove(id: number) {
-    return this.prisma.product.delete({ where: { id } });
+  async remove(id: number): Promise<void> {
+    const product = await this.prisma.product.findUnique({ where: { id } });
+    if (!product) {
+      throw new NotFoundException(`Product with id ${id} does not exist`);
+    }
+    await this.prisma.product.delete({ where: { id } });
   }
 }
